# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transporte API",
        "description": "Documentación generada con Swagger para la API de transporte",
        "contact": {
            "email": "soporte@transporte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://tu-dominio.com",
            "description": "Servidor Producción"
        }
    ],
    "paths": {
...
        "/api/detalles-paquetes": {
            "get": {
                "tags": [
                    "Detalles de Paquetes"
                ],
                "summary": "Listar detalles de paquetes",
                "description": "Obtiene un listado paginado de todos los detalles de paquetes junto con su paquete y tipo de mercancía.",
                "operationId": "b07f52da1ebda2442d0e242a50a042da",
                "responses": {
                    "200": {
                        "description": "Lista de detalles de paquetes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DetallePaquete"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Detalles de Paquetes"
                ],
                "summary": "Crear un nuevo detalle de paquete",
                "operationId": "690a7893b3c1888584afcc669c13c167",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paquete_id",
                                    "tipo_mercancia_id",
                                    "dimension",
                                    "peso",
                                    "fecha_entrega"
                                ],
                                "properties": {
                                    "paquete_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_mercancia_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "dimension": {
                                        "type": "string",
                                        "example": "50x40x30 cm"
                                    },
                                    "peso": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "fecha_entrega": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detalle de paquete creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetallePaquete"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/detalles-paquetes/{id}": {
            "get": {
                "tags": [
                    "Detalles de Paquetes"
                ],
                "summary": "Mostrar un detalle de paquete",
                "description": "Devuelve la información de un detalle de paquete específico junto con su paquete y tipo de mercancía.",
                "operationId": "7888d7b42d0ddb3c33c6655fc8be8b02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de paquete encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetallePaquete"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Detalles de Paquetes"
                ],
                "summary": "Actualizar un detalle de paquete",
                "operationId": "167e9bae16d51dbbe186e290c25596cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paquete_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_mercancia_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "dimension": {
                                        "type": "string",
                                        "example": "60x45x35 cm"
                                    },
                                    "peso": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 30.2
                                    },
                                    "fecha_entrega": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-11-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalle de paquete actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetallePaquete"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Detalles de Paquetes"
                ],
                "summary": "Eliminar un detalle de paquete",
                "operationId": "5ab4dc10dcc959ebd27a3c5c8c8d84d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de paquete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de paquete eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detalle de paquete eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DetallePaquete": {
                "title": "DetallePaquete",
                "description": "Detalles específicos de un paquete",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "paquete_id": {
                        "description": "ID del paquete asociado",
                        "type": "integer",
                        "example": 10
                    },
                    "tipo_mercancia_id": {
                        "description": "ID del tipo de mercancía",
                        "type": "integer",
                        "example": 2
                    },
                    "dimension": {
                        "description": "Dimensiones del paquete (largo x ancho x alto)",
                        "type": "string",
                        "example": "40x30x20 cm"
                    },
                    "peso": {
                        "description": "Peso del paquete en kilogramos",
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "fecha_entrega": {
                        "description": "Fecha estimada de entrega del paquete",
                        "type": "string",
                        "format": "date",
                        "example": "2025-10-15"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}